backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
Pillow==10.4.0
torch==2.3.1+cpu
torchvision==0.18.1+cpu
pydantic==2.8.2
